Feature: Add dynamic worker pool

+++ internal/app/usecases/downloader/execute_download_task.go
@@ -16,7 +16,7 @@ import (
 
 func (uc *YouTubeDownloader) ExecuteDownloadTask(
 	ctx context.Context,
-	workerId uint,
+	workerId uint64,
 	task *ddownload.DownloadTask,
 ) error {
 	if task == nil {
