+++ internal/ports/workers/download_task_runner.go
@@ -7,5 +7,5 @@ import (
 )
 
 type DownloadTaskRunner interface {
-	ExecuteDownloadTask(ctx context.Context, workerId uint, task *ddownload.DownloadTask) error
+	ExecuteDownloadTask(ctx context.Context, workerId uint64, task *ddownload.DownloadTask) error
 }
