+++ pkg/nworkerpool/job.go
@@ -8,7 +8,7 @@ import (
 type Job interface {
 	ID() string
 	Name() string
-	Execute(ctx context.Context, workerID uint) error
+	Execute(ctx context.Context, workerID uint64) error
 }
 
 type JobDispatcher interface {
@@ -27,13 +27,15 @@ type task struct {
 	job    Job
 }
 
-func newJobQueue(cap int) *jobQueue {
+// newJobQueue creates a new job queue with the specified capacity.
+func newJobQueue(cap uint32) *jobQueue {
 	return &jobQueue{
 		items: list.New(),
 		index: make(map[string]*list.Element, cap),
 	}
 }
 
+// newTask creates a new task with the specified context and job.
 func newTask(ctx context.Context, job Job) *task {
 	ctx, cancel := context.WithCancel(ctx)
 
@@ -44,10 +46,13 @@ func newTask(ctx context.Context, job Job) *task {
 	}
 }
 
+// Len returns the number of items in the queue.
 func (q *jobQueue) Len() int {
 	return q.items.Len()
 }
 
+// Push adds a new item to the queue.
+// If an item with the same ID already exists, it is not added and false is returned.
 func (q *jobQueue) Push(job Job) bool {
 	id := job.ID()
 	if _, exists := q.index[id]; exists {
@@ -59,6 +64,8 @@ func (q *jobQueue) Push(job Job) bool {
 	return true
 }
 
+// Pop removes and returns the first item in the queue.
+// If the queue is empty, it returns nil and false.
 func (q *jobQueue) Pop() (Job, bool) {
 	el := q.items.Front()
 	if el == nil {
@@ -74,6 +81,8 @@ func (q *jobQueue) Pop() (Job, bool) {
 	return job, true
 }
 
+// Remove removes an item from the queue by ID.
+// If no such item exists, it returns false.
 func (q *jobQueue) Remove(id string) bool {
 	el, exists := q.index[id]
 	if !exists {
